requirements.txt
pandas
requests
matplotlib
seaborn
jupyter
python-dotenv